+++ pakyow-core/spec/unit/app/instance_handle_spec.rb
@@ -1,0 +1,11 @@
+RSpec.describe Pakyow::Application, "#handle" do
+  let(:instance) {
+    described_class.new
+  }
+
+  it "is not defined" do
+    expect {
+      instance.handle
+    }.to raise_error(NoMethodError)
+  end
+end